fix: corrections for Postman Main/Stat tests

+++ main-service/src/main/java/ewm/event/controller/EventAdminController.java
@@ -23,7 +23,7 @@ public class EventAdminController {
     @ResponseStatus(HttpStatus.OK)
     public List<EventFullDto> get(@RequestParam(required = false) List<Long> users,
                                   @RequestParam(required = false) List<EventState> states,
-                                  @RequestParam(required = false) List<Integer> categories,
+                                  @RequestParam(required = false) List<Long> categories,
                                   @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
                                   LocalDateTime rangeStart,
                                   @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
